+++ src/lib/StaticTypeaheadInput.tsx
@@ -3,16 +3,14 @@ import { Controller, FieldValues, useFormContext } from "react-hook-form";
 import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
 import { FormGroupLayout } from "./FormGroupLayout";
 import { convertTypeaheadOptionsToStringArray } from "./helpers/typeahead";
-import { CommonInputProps } from "./types/CommonInputProps";
-import { TypeaheadOptions } from "./types/Typeahead";
+import { CommonTypeaheadProps, TypeaheadOptions } from "./types/Typeahead";
 
-interface StaticTypeaheadInputProps<T extends FieldValues> extends CommonInputProps<T> {
-  multiple?: boolean;
+interface StaticTypeaheadInputProps<T extends FieldValues> extends CommonTypeaheadProps<T> {
   options: TypeaheadOptions;
 }
 
 const StaticTypeaheadInput = <T extends FieldValues>(props: StaticTypeaheadInputProps<T>) => {
-  const { disabled, label, helpText, labelToolTip } = props;
+  const { disabled, label, helpText, labelToolTip, defaultSelected } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
 
   const { control } = useFormContext();
@@ -25,6 +23,7 @@ const StaticTypeaheadInput = <T extends FieldValues>(props: StaticTypeaheadInput
         <FormGroupLayout helpText={helpText} name={name} id={id} label={label} labelToolTip={labelToolTip}>
           <Typeahead
             {...field}
+            defaultSelected={defaultSelected}
             multiple={props.multiple}
             onChange={(e) => {
               const values = convertTypeaheadOptionsToStringArray(e);